Slight change to MeasureOverlap for binary vectors. OLD: score = 1 - normalized hamming distance NEW: score = 2 * (0.5 - normalized hamming distance)
The idea being that a normalized hamming distance of 0.5 is equivalent to orthogonality, so it's misleading to include this anticipated random overlap in the similarity metric. 

Also have requested that the VectorSearcherSubspaceSim not attempt to orthogonalize binary vectors, at least until such time as we include something analogous to orthogonalization for these vectors.

The combination of these two changes produces reasonable results for subspace-like searches with binary vectors, assuming the component vectors of the subspace are dissimilar (normHD +- 0.5) to start out with.