Added support for Enums to FlagConfig. No more VectorType.valueOf() - w00t!
As a consequence, header strings in vector stores will say (e.g.) BINARY instead of binary, but reading code is robust to case so I decided to allow this as a (I think) backward-compatible change.

This paves the way for term weighting, searchtype, and any other enums to be parsed on the way in. This is good - error checking is early, raw strings aren't passed around and parsed later.

Tests all pass.